Simplify lazy loading of readers
