--- neurom/io/tests/test_h5_mock.py
@@ -1,47 +1,0 @@
-# Copyright (c) 2015, Ecole Polytechnique Federale de Lausanne, Blue Brain Project
-# All rights reserved.
-#
-# This file is part of NeuroM <https://github.com/BlueBrain/NeuroM>
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-#     1. Redistributions of source code must retain the above copyright
-#        notice, this list of conditions and the following disclaimer.
-#     2. Redistributions in binary form must reproduce the above copyright
-#        notice, this list of conditions and the following disclaimer in the
-#        documentation and/or other materials provided with the distribution.
-#     3. Neither the name of the copyright holder nor the names of
-#        its contributors may be used to endorse or promote products
-#        derived from this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY
-# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-import os
-from neurom.io._hdf5mock import H5
-from nose import tools as nt
-
-_path = os.path.dirname(os.path.abspath(__file__))
-DATA_PATH = os.path.join(_path, '../../../test_data')
-H5_PATH = os.path.join(DATA_PATH, 'h5')
-H5V1_PATH = os.path.join(H5_PATH, 'v1')
-H5V2_PATH = os.path.join(H5_PATH, 'v2')
-
-
-@nt.raises(ImportError)
-def test_read_h5_v1_raises():
-    H5.read(os.path.join(H5V1_PATH, 'Neuron.h5'))
-
-
-@nt.raises(ImportError)
-def test_read_h5_v2_raises():
-    H5.read(os.path.join(H5V2_PATH, 'Neuron.h5'))
